Image slider move to component
