+++ components/imgSlider/imgSlider.module.css
@@ -1,0 +1,5 @@
+.image {
+  object-fit: cover;
+  width: 100%;
+  height: auto;
+}
